Fixed bug in the related-product-list where image doesn't match description
