+++ package.json
@@ -25,6 +25,7 @@
     "express": "^4.17.1",
     "express-http-proxy": "^1.6.2",
     "jest": "^26.6.3",
+    "jquery": "^3.6.0",
     "nodemon": "^2.0.7",
     "path": "^0.12.7",
     "proxy": "^1.0.2",
